extract generic auth objects into module
