+++ jobserver/settings.py
@@ -266,7 +266,7 @@ SERVER_EMAIL = "your-server@example.com"
 # https://www.django-rest-framework.org/api-guide/settings/
 REST_FRAMEWORK = {
     "DEFAULT_AUTHENTICATION_CLASSES": [
-        "jobserver.api.NoAuthentication",
+        "jobserver.api.authentication.NoAuthentication",
     ],
     "DEFAULT_PAGINATION_CLASS": "rest_framework.pagination.LimitOffsetPagination",
     "PAGE_SIZE": 100,
